tests/wpt/web-platform-tests/bluetooth/requestDevice/acceptAllDevices/device-with-empty-name.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setUpPreconnectedDevice({name: ''})
    .then(() => requestDeviceWithTrustedClick({acceptAllDevices: true}))
    .then(device => {
      assert_equals(device.name, '');
    });
}, 'Device with empty name and no UUIDs nearby. Should be found if ' +
   'acceptAllDevices is true.');
</script>
